Set remote_host to first IP in X-Forwarded-For header
When running sstp-server behind an HTTP (reverse) proxy, the IP
address/host determined from the transport (info) peername is the one
from the proxy server, rather than from the original peer.

An HTTP reverse proxy may forward the (list of) IP addresses of a peer
connection in the X-Forwarded-For HTTP header. If so detected by
sstp-server, set the remote_host to the first (originator) IP address
from that list.

Signed-off-by: Tijs Van Buggenhout <tijs.van.buggenhout@axsguard.com>