Terminate child PPP process instead of closing pipe
Avoids that IP addresses allocated from the pool are not properly
unregistered when the client connection is terminated.

Calling close on an already closed transport, ie it may be that the PPP
process is already gone, and thus the (transport) pipe already closed,
will raise an exception, and cause that the client's internal IP is not
unregistered in the pool.

By instead terminating the PPP process, and handling any possible
exception related, the pipe will be closed automatically and the
client's internal IP address is always unregistered.

Signed-off-by: Tijs Van Buggenhout <tijs.van.buggenhout@axsguard.com>