Integrate the PPP SSTP API into sstp-server
Detects the presence of the sstp-pppd-plugin.so plugin on the system,
and creates a task to start communication over a unix socket of the
PPP SSTP API protocol if present.

It allows for ppp process to pass the higher layer authentication keys
to the sstp-server instance in order to verify the crypto binding.

If the plugin is not available, crypto binding verification will not be
supported.

Signed-off-by: Tijs Van Buggenhout <tijs.van.buggenhout@axsguard.com>