Verify crypto binding set by client using HLAK
Implements the Crypto Binding Compound MAC (CMAC) verification,
when the Crypto Binding Attribute is present in the SSTP Call Connected
message.

The PPP (sub)process passes the MPPE send/receive keys when CHAP
authentication succeeds. When received by sstp-server, it installs them
as the Higher Layer Key (HLAK). In server mode the HLAK is obtained by
concatenating the master receive key with the send key. Ensure the key
is 32 bytes long (should be the case for MPPE) - zero-padded if needed.

The Crypto Binding CMAC is (re)calculated on server side and compared
with the value present in the Crypto Binding Attribute. In case they do
not match the connection is aborted.

CMAC uses HMAC using the Crypto Binding Hash algorithm negotiated (SHA1
or SHA256). Input for HMAC is the complete SSTP Call Connected message,
with CMAC field in Crypto Binding Attribute zeroed out. The key for the
HMAC operation is obtained using the PRF algorithm based on PRF+ from
IKEv2, called the Compound Mac Key (CMK).

CMK is obtained from the first iteration of PRF using the same HMAC
cryptographic operation. Input for CMK is the CMK seed 'SSTP inner
method derived CMK' concatenated with the little endian formatted length
of the cryptographic hash used and the iteration number, whereas the key
is HLAK.

In order to support sstp-server without sstp-pppd-plugin.so, CMAC is not
verified when plugin is not available.

Resolves sorz/sstp-server#1

Signed-off-by: Tijs Van Buggenhout <tijs.van.buggenhout@axsguard.com>