Restrict X-Forwarded-For processing when to 'no_ssl'
Only process the X-Forwarded-For HTTP header if set by a trusted
component, e.g. in a reverse proxy.

Since 'no_ssl' is only meaningful behind an (SSL) reverse proxy, it is a
good candidate to enable processing of the header.

Note that (SSL) reverse proxy to SSL-enabled sstpd currently does not
enable processing of the header.

Signed-off-by: Tijs Van Buggenhout <tijs.van.buggenhout@axsguard.com>