Set remote_host to first IP in X-Forwarded-For header (#36)
* Set remote_host to first IP in X-Forwarded-For header

When running sstp-server behind an HTTP (reverse) proxy, the IP
address/host determined from the transport (info) peername is the one
from the proxy server, rather than from the original peer.

An HTTP reverse proxy may forward the (list of) IP addresses of a peer
connection in the X-Forwarded-For HTTP header. If so detected by
sstp-server, set the remote_host to the first (originator) IP address
from that list.

Signed-off-by: Tijs Van Buggenhout <tijs.van.buggenhout@axsguard.com>

* Restrict X-Forwarded-For processing when to 'no_ssl'

Only process the X-Forwarded-For HTTP header if set by a trusted
component, e.g. in a reverse proxy.

Since 'no_ssl' is only meaningful behind an (SSL) reverse proxy, it is a
good candidate to enable processing of the header.

Note that (SSL) reverse proxy to SSL-enabled sstpd currently does not
enable processing of the header.

Signed-off-by: Tijs Van Buggenhout <tijs.van.buggenhout@axsguard.com>