Override remote_port from X-Forwarded-SourcePort
When sstp-server is used behind an HTTP reverse proxy, the remote IP
from the client is passed by the X-Forwarded-For header. The same can be
done for the remote port number, using the X-Forwarded-SourePort header.

In case only the X-Forwarded-For header is passed, the remote_port (set
from the peer transport information) is reset to None, as this port
number is by the HTTP reverse proxy, rather than the SSTP client.

Signed-off-by: Tijs Van Buggenhout <tijs.van.buggenhout@axsguard.com>