Support empty/zero HLAK in crypto biding validation
When a higher layer authentication key is not available, or it is
bypassed, an empty/zero HLAK (32 zero octets) must be used to calculate
the crypto binding [1].

Since the server can not know beforehand whether or not this was the case
for the client, the two - zero HLAK and server HLAK if available - need
to be considered, and one successful validation is sufficient to accept
the crypto binding.

At least for Windows10 clients it seems to be the case that for normal
MS-CHAPv2 authentication HLAK is available, whereas it is not for
EAP-MS-CHAPv2. The specification [1] desribes MS-CHAPv2, EAP-TLS, and
other EAP methods. EAP-MSCHAP-V2 falls under the latter and is reffered
to RFC-3748 [2], where it is stated that it may be possible for peer and
server to mutually authenticate and derive keys. If so a (extended)
master session key of at least 64 octets must be derived. However the
EAP-MS-CHAPv2 specification [3] does not mention key derivation.

According to VPN authentication options documentation [4],
EAP-PEAP-MS-CHAPv2 should provide key derivation and a master session
key. It even provides a crypto-binding TLV.

1. [MS-SSTP] sections 3.2.5.2.{2,4}
2. https://www.ietf.org/rfc/rfc3748.txt section 7.10
3. https://tools.ietf.org/html/draft-kamath-pppext-eap-mschapv2-02
4. https://docs.microsoft.com/en-us/windows/security/identity-protection/vpn/vpn-authentication

Signed-off-by: Tijs Van Buggenhout <tijs.van.buggenhout@axsguard.com>