Fix decoding of SSTP-API messages.
libsstp-api has a really funny way of writing its messages and the current
code in sstp-server breaks on big-endian machines. In addition, attribute
data is encoded into 32-bit blocks. So, say, a 1-byte value will actually
occupy 4 bytes on the "wire". The current code in sstp-server breaks on
this as well.