Fix exceptions in rump connections.
In the Wild West that is the public Internet, you sometimes get a connection
that executes all or a part of the TLS handshake and then does nothing or
drops the connection. sstp-server starts a Hello timer as soon as a
connection is established, and it will expire some time after the client
drops the connection. Since the SSTPProtocol object doesn't have the
transport attribute when the Hello timer expires, an exception is generated.

BTW, a review of code should be performed since it's possible there's a
memory leak in these or other pathologic situations.